feat: create useRankingProducts custom hook
